Move ThemeResourceKeys to Resourcing
